refactor: move command methods info command file
