+++ Cargo.toml
@@ -18,6 +18,7 @@ log = { version = "0" }
 env_logger = "0"
 futures = "0"
 walkdir = "2"
+async-trait = "0.1.73"
 
 # ignore = "0"
 # xattr = "0"
