+++ src/constants.rs
@@ -1,3 +1,5 @@
 pub(crate) const PACK_NAME_ENV: &str = "PACK_NAME";
 
 pub(crate) const CONFIG_FILE_NAME: &str = "stow-cm.toml";
+
+pub(crate) const GLOBAL_CONFIG_FILE: &str = "${XDG_CONFIG_HOME:-~/.config}/stow-cm/config.toml";
